requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
discord.py==2.0.1
distlib==0.3.6
filelock==3.8.0
frozenlist==1.3.1
idna==3.4
jokeapi==1.0.5
MouseInfo==0.1.3
multidict==6.0.2
Pillow==9.3.0
platformdirs==2.5.4
prawcore==2.3.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dotenv==0.21.0
pytweening==1.0.4
requests==2.28.1
simplejson==3.18.1
update-checker==0.18.0
urllib3==1.26.13
virtualenv==20.16.7
wavelink==1.3.3
websocket-client==1.4.2
yarl==1.8.1
